Fix create or update multiple choice property

+++ digiroad2-oracle/src/main/scala/fi/liikennevirasto/digiroad2/asset/oracle/MassTransitStopDao.scala
@@ -177,7 +177,7 @@ class MassTransitStopDao {
   }
 
   private[this] def createOrUpdateMultipleChoiceProperty(propertyValues: Seq[PropertyValue], assetId: Long, propertyId: Long) {
-    val newValues = propertyValues.map(_.propertyValue)
+    val newValues = propertyValues.map(_.propertyValue.toLong)
     val currentIdsAndValues = Q.query[(Long, Long), (Long, Long)](multipleChoicePropertyValuesByAssetIdAndPropertyId).apply(assetId, propertyId).list
     val currentValues = currentIdsAndValues.map(_._2)
     // remove values as necessary
@@ -192,7 +192,7 @@ class MassTransitStopDao {
       !currentValues.contains(_)
     }.foreach {
       v =>
-        insertMultipleChoiceValue(assetId, propertyId, v.toLong).execute
+        insertMultipleChoiceValue(assetId, propertyId, v).execute
     }
   }
